Add Typed SPI and refactor related SPI
